feat: Add Finger and Membership models
This commit adds the `Finger` and `MembershipTypeEnum` models to the `client` module. The `Finger` model represents a fingerprint with `fingerprint` and `tmp` attributes. The `MembershipTypeEnum` model is an enumeration with two values: `IS_MEMBER` and `NOT_MEMBER`. These models are necessary for future functionality related to client data.

chore: Add User and ModifyUser models

This commit introduces the `User` and `ModifyUser` models to the `user` module. The `User` model represents a user with attributes such as `first_name`, `last_name`, `email`, and `password`. The `ModifyUser` model is used for modifying user data and has optional attributes. These models are essential for user management features.

chore: Update Finger create controller

This commit updates the `create_worker_finger` function in the `finger/create.py` controller. It removes the unused `BackgroundTasks` import and the `json` import. This change improves the efficiency and readability of the code.

chore: Add create user controller

This commit adds the `create_user` function to the `users/create.py` controller. This function handles the creation of new users and includes necessary imports and dependencies.

chore: Add database configuration and utility functions

The code changes include:
- Adding a `.env.example` file with placeholders for database configuration variables
- Creating a `db.py` module in the `app/services` directory to handle database connections
- Implementing a `Env` class in the `env.py` module to retrieve environment variables securely
- Adding a `get_finger_image` function in the `get.py` module to fetch fingerprint images from the database
- Updating the `main.py` file to include a home route that streams data
- Creating a `fastapi.py` module in the `app/services` directory to configure the FastAPI application
- Adding dependencies to the `requirements.txt` file
- Creating a `login.py` module in the `app/utils` directory to verify user passwords